todo_test/requirements.txt
# Requirements
# Python 3.9.7
djangorestframework==3.12.4